# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "PublicUsersController_getPublicProfile",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–50, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Target user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public profile and products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public profile and active listings for a user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/ratings": {
      "post": {
        "operationId": "RatingsController_rateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating submitted or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot rate yourself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rate a user (1–5 stars)",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/ratings/{userId}": {
      "get": {
        "operationId": "RatingsController_getUserRatingSummary",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Target user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Average rating and review count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingSummaryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get rating summary for a user",
        "tags": [
          "Ratings"
        ]
      }
    },
...
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user details for moderation page (admin only)",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user (soft-delete, admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/v2/ratings": {
      "post": {
        "operationId": "KayanController_createItemRating",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KayanRatingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Kayan V2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Place API",
    "description": "Online marketplace REST API — user registration, product listings, real-time chat, ratings, and admin moderation.\n\n> **WebSocket (Chat):** Real-time messaging uses a Socket.io gateway at the `/chat` namespace. That interface is not documented here; see the project README for event names and payloads.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "example": null,
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "RelatedFileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 10
          },
          "purpose": {
            "type": "string",
            "example": "product_image",
            "enum": [
              "avatar",
              "product_image",
              "chat_attachment",
              "document"
            ]
          },
          "object_key": {
            "type": "string",
            "example": "products/91/1.jpg"
          },
          "mime_type": {
            "type": "object",
            "example": "image/jpeg",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "uploaded",
            "enum": [
              "pending",
              "uploaded",
              "failed",
              "deleted"
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "uploaded_at": {
            "type": "object",
            "example": "2026-03-28T12:10:00.000Z",
            "nullable": true
          },
          "url": {
            "type": "string",
            "example": "https://res.cloudinary.com/example/image/upload/products/91/1.jpg"
          }
        },
        "required": [
          "id",
          "purpose",
          "object_key",
          "status",
          "created_at",
          "url"
        ]
      },
...
      "SuccessDataDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/SuccessDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "name": {
            "type": "string",
            "example": "Jana Ahmed"
          },
          "member_since": {
            "type": "string",
            "example": "2025-02-22T10:00:00.000Z"
          },
          "ads_count": {
            "type": "number",
            "example": 10
          },
          "rate": {
            "type": "string",
            "example": "4.50"
          },
          "avatar": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "contactInfo": {
            "type": "object",
            "example": "+201000000001",
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "id",
          "name",
          "member_since",
          "ads_count",
          "rate"
        ]
      },
      "RelatedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "ssn": {
            "type": "object",
            "example": "29876543210987",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Jana Ahmed"
          },
          "phone": {
            "type": "string",
            "example": "+201000000012"
          },
          "profileState": {
            "type": "string",
            "example": "active",
            "nullable": true,
            "enum": [
              "active",
              "paused",
              "banned"
            ]
          },
          "rate": {
            "type": "object",
            "example": "4.50",
            "nullable": true
          },
          "avatar": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "contactInfo": {
            "type": "object",
            "example": "+201000000012",
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "id",
          "name",
          "phone"
        ]
      },
      "ProductImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "file": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "sort_order": {
            "type": "number",
            "example": 0
          },
          "object_key": {
            "type": "string",
            "example": "products/1/image.jpg"
          },
          "status": {
            "type": "string",
            "example": "uploaded",
            "enum": [
              "pending",
              "uploaded",
              "failed",
              "deleted"
            ]
          }
        },
        "required": [
          "id",
          "file",
          "sort_order",
          "object_key",
          "status"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "owner": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "subcategory": {
            "type": "object",
            "example": "Mobile Phones",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Used Laptop"
          },
          "description": {
            "type": "object",
            "example": "Good condition, 1 year old",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 1500
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "address_text": {
            "type": "object",
            "example": "10 Tahrir Square",
            "nullable": true
          },
          "details": {
            "type": "object",
            "example": {
              "condition": "used",
              "brand": "Apple",
              "storage": "256GB"
            },
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "available",
            "enum": [
              "available",
              "sold",
              "archived"
            ]
          },
          "is_negotiable": {
            "type": "boolean",
            "example": false
          },
          "preferred_contact_method": {
            "type": "string",
            "example": "both",
            "enum": [
              "phone",
              "chat",
              "both"
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "seller_rate": {
            "type": "object",
            "example": "4.50",
            "description": "Seller average rating — only present in search results",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          }
        },
        "required": [
          "id",
          "owner",
          "category",
          "name",
          "price",
          "city",
          "status",
          "is_negotiable",
          "preferred_contact_method",
          "created_at",
          "updated_at",
          "images"
        ]
      },
      "PublicUserProfileDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        },
        "required": [
          "user",
          "products"
        ]
      },
      "PublicUserProfileResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/PublicUserProfileDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateRatingDto": {
        "type": "object",
        "properties": {
          "ratedUserId": {
            "type": "number",
            "description": "ID of the user being rated",
            "example": 15,
            "minimum": 1
          },
          "ratingValue": {
            "type": "number",
            "description": "Star rating (1–5)",
            "example": 4,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Optional review comment (1–2000 chars)",
            "example": "Great seller, fast shipping!",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "ratedUserId",
          "ratingValue"
        ]
      },
      "RatingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "rater": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "rated_user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "rating_value": {
            "type": "number",
            "example": 4,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "object",
            "example": "Great seller, fast delivery",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "rater",
          "rated_user",
          "rating_value",
          "created_at",
          "updated_at"
        ]
      },
      "RatingDataDto": {
        "type": "object",
        "properties": {
          "rating": {
            "$ref": "#/components/schemas/RatingDto"
          }
        },
        "required": [
          "rating"
        ]
      },
      "RatingResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/RatingDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "RatingSummaryDto": {
        "type": "object",
        "properties": {
          "avg_rating": {
            "type": "string",
            "example": "4.50",
            "description": "Average rating (2 decimal places as numeric string)"
          },
          "ratings_count": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "avg_rating",
          "ratings_count"
        ]
      },
      "RatingSummaryDataDto": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/RatingSummaryDto"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingDto"
            }
          }
        },
        "required": [
          "summary",
          "ratings"
        ]
      },
      "RatingSummaryResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/RatingSummaryDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "AdminUserDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "ssn": {
            "type": "object",
            "example": "29876543210987",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Jana Ahmed"
          },
          "phone": {
            "type": "string",
            "example": "+201000000012"
          },
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "paused",
              "banned"
            ]
          },
          "profileState": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "paused",
              "banned"
            ]
          },
          "avatar": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "contactInfo": {
            "type": "object",
            "example": "+201000000012",
            "nullable": true
          },
          "is_admin": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "status",
          "profileState",
          "is_admin",
          "created_at",
          "updated_at"
        ]
      },
      "AdminUserDetailsDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AdminUserDetailsDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "AdminUserDetailsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserDetailsDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateItemRatingDto": {
        "type": "object",
        "properties": {}
      },
      "ItemRatingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "item_type": {
            "type": "object",
            "nullable": true
          },
          "item_id": {
            "type": "number",
            "nullable": true
          },
          "order_id": {
            "type": "number",
            "nullable": true
          },
          "product_id": {
            "type": "number",
            "nullable": true
          },
          "rating_value": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "user_id",
          "rating_value"
        ]
      },
      "KayanRatingDataDto": {
        "type": "object",
        "properties": {
          "rating": {
            "$ref": "#/components/schemas/ItemRatingDto"
          }
        },
        "required": [
          "rating"
        ]
      },
      "KayanRatingResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/KayanRatingDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}